fix(emit): order private helpers after invalid static reads
